refactor: select로직 상위 컴포넌트로 분리
